add missing documentation for datatypes
